rename kotlin app to LevelPlay Demo
