+++ Android/Kotlin/app/build.gradle
@@ -14,11 +14,11 @@ ext {
 }
 
 android {
-    namespace "com.levelplay.levelplaysdkdemo"
+    namespace "com.levelplay.levelplaydemo"
     compileSdkVersion 34
     buildToolsVersion '34.0.0'
     defaultConfig {
-        applicationId "com.levelplay.levelplaysdkdemo"
+        applicationId "com.levelplay.levelplaydemo"
         minSdkVersion 19
         targetSdkVersion 34
         versionCode 1
